page number limit

+++ lib/workstation.sh
@@ -150,7 +150,9 @@ workstation::backup_my_github_repositories() {
 
   local full_name
   local git_url
-  local page_number=1
+
+  # NOTE: There is a 100 000 (1000*100) repository limit here. I put it here to not suffer an infinite loop if something is wrong
+  local page_number_limit=1000
 
   if [ -t 0 ]; then # stdin is a terminal
     local fail_command="fail"
@@ -159,8 +161,8 @@ workstation::backup_my_github_repositories() {
   fi
 
   dir::make_if_not_exists "${backup_path}" || fail
-
-  while true; do
+  
+  local page_number; for ((page_number=1; page_number<=page_number_limit; page_number++)); do
     # url for public repos for the specific user "https://api.github.com/users/${MY_GITHUB_LOGIN}/repos?page=${page_number}&per_page=100"
     curl \
       --fail \
@@ -185,9 +187,11 @@ workstation::backup_my_github_repositories() {
     elif [ "${saved_pipe_status[*]}" != "0 0 0" ]; then
       fail "Abnormal termination: ${saved_pipe_status[*]}"
     fi
-
-    ((page_number++))
   done
+
+  if [ "$page_number" -gt "$page_number_limit" ]; then
+    fail "page number limit reached"
+  fi
 
   if [ -f "${fail_flag}" ]; then
     rm "${fail_flag}" || fail
